Move more to utils and provenance.py
