+++ cohorts/__init__.py
@@ -15,6 +15,7 @@
 from .cohort import Cohort
 from .patient import Patient
 from .sample import Sample
+from .dataframe_loader import DataFrameLoader
 
 from ._version import get_versions
 __version__ = get_versions()['version']
